+++ classes/TextClass/Text.cpp
@@ -19,7 +19,7 @@ Bounds Text::getBounds() const
 	return bounds;
 }
 
-void Text::setPosition(double x, double y)
+void Text::setPosition(float x, float y)
 {
 	bounds.left	  = x;
 	bounds.right  = x + bounds.right - bounds.left;
